requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

agate==1.6.3
asn1crypto==1.4.0
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
dbt-core==1.0.0
dbt-extractor==0.4.0
dbt-postgres==1.0.0
dbt-snowflake==1.0.0
future==0.18.2
hologram==0.0.14
idna==3.3
importlib-metadata==4.11.1
isodate==0.6.1
Jinja2==2.11.3
jsonschema==3.1.1
keyring==23.5.0
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.0.1
mashumaro==2.9
minimal-snowplow-tracker==0.0.2
msgpack==1.0.3
networkx==2.6.3
oscrypto==1.2.1
packaging==21.3
parsedatetime==2.4
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodomex==3.14.1
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-slugify==6.0.1
pytimeparse==1.1.8
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
snowflake-connector-python==2.7.4
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.7.0
google-api-core==1.16.0
googleapis-common-protos==1.6.0
azure-functions
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
azure-storage-queue==1.4.0
opencensus>=0.7.6
opencensus-context>=0.1.1
opencensus-ext-azure>=1.0.1
pytest>=5.3.5
schema>=0.7.1










